Cwdoe 564 drag and drop reorder msel events (#151)
This MR adds the ability to drag and drop scenario events on the Scenario Events page to reorder them. Reordering will attempt to move the events before, after, or between the adjacent events, if any, if there is a separation of DeltaSeconds among them. If not, then it updates the GroupOrder for events that happen at the same time.

Because sorting and filtering logic will also be implemented for the Exercise View page, the MR abstracts that functionality out of the ScenarioEvents page into the ScenarioEventDataService so that the logic can be shared. Some other functionality is also moved there because it is used an several other pages where we have duplicated similar code. The intent is to clean up those places in subsequent tickets.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>