Added ability for player to control the theme of the embedded event management page (#158)
This MR adds the {theme} parameter to the event management page so that player can manage its theme when it is embedded. Also, removed the redundant ui selected theme setting because that preference is now saved as part of the akita datastore cache. Removed top component from all pages when embedded in another app like player since the containing app should own that menu.